refactor(types): consolidate file reference types to unified location
- Add createFileReference function to file-transfer/types.ts for backward compatibility
- Convert src/types/file-reference.ts to a re-export module with @deprecated notice
- All file-related types are now consolidated in src/file-transfer/types.ts
- Maintains full backward compatibility through re-exports

Related: #194
